refactor(appsec): move normalizeHeaders
